SilentSignals project fixed

+++ src/main/java/com/silentsignals/silentsignals/controller/ContactController.java
@@ -1,10 +1,14 @@
 package com.silentsignals.silentsignals.controller;
 
+import com.silentsignals.silentsignals.dto.ContactRequestDto;
 import com.silentsignals.silentsignals.entity.Contact;
 import com.silentsignals.silentsignals.entity.User;
 import com.silentsignals.silentsignals.repository.ContactRepository;
 import com.silentsignals.silentsignals.repository.UserRepository;
+import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,21 +19,40 @@ public class ContactController {
 
     private final ContactRepository contactRepository;
     private final UserRepository userRepository;
+
+    @PreAuthorize("hasAnyAuthority('ROLE_USER','ROLE_ADMIN')")
     @PostMapping
-    public Contact add(@RequestBody Contact c) {
-        if (c.getUser() != null && c.getUser().getId() != null) {
-            User user = userRepository.findById(c.getUser().getId())
-                    .orElseThrow(() -> new RuntimeException("User tapılmadı"));
-            c.setUser(user);
-        }
+    public Contact add(@Valid @RequestBody ContactRequestDto dto) {
+
+        User user = userRepository.findById(dto.getUserId())
+                .orElseThrow(() -> new RuntimeException("User tapılmadı"));
+
+        Contact c = new Contact();
+        c.setUser(user);
+        c.setName(dto.getName());
+        c.setEmail(dto.getEmail());
+        c.setPhone(dto.getPhone());
+
         return contactRepository.save(c);
     }
+
+    @PreAuthorize("hasAnyAuthority('ROLE_USER','ROLE_ADMIN')")
     @GetMapping
-    public List<Contact> all() {
-        return contactRepository.findAll();
+    public List<Contact> all(Authentication auth) {
+        return contactRepository.findByUserUsername(auth.getName());
     }
+
+    @PreAuthorize("hasAnyAuthority('ROLE_USER','ROLE_ADMIN')")
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable Long id) {
+    public void delete(@PathVariable Long id, Authentication auth) {
+
+        Contact contact = contactRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Contact tapılmadı"));
+
+        if (!contact.getUser().getUsername().equals(auth.getName())) {
+            throw new RuntimeException("Access denied");
+        }
+
         contactRepository.deleteById(id);
     }
 }
